feat: adding basic autocompletion to language
